CLN: Move boxing logic to BlockManager
